refactor(api): extract shared RAG domain entities into core/rag/entities/
fix

fix

refactor(api): extract shared RAG domain entities into core/rag/entities/

fix

fix

fix
